+++ nautobot_device_lifecycle_mgmt/jobs/__init__.py
@@ -1,0 +1,5 @@
+"""Nautobot Jobs for the Device Lifecycle plugin."""
+from .cve_tracking import GenerateVulnerabilities
+from .lifecycle_reporting import DeviceSoftwareValidationFullReport, InventoryItemSoftwareValidationFullReport
+
+jobs = [DeviceSoftwareValidationFullReport, InventoryItemSoftwareValidationFullReport, GenerateVulnerabilities]
